feat(grz-gatekeeper): refactor into modules, rate limiting
